refactor: change inertial manager class
